extract Behinds in standalone classes @Layout
