+++ src/Witness/NTS.Witness.Web/Components/Layout/MainLayout.razor
@@ -1,7 +1,7 @@
 ﻿@using Microsoft.AspNetCore.Components.Authorization
 @using Not.Blazor.Notifier
 
-@inherits LayoutComponentBase
+@inherits MainLayoutBehind
 
 <NTheme Mobile="true"/>
 
@@ -12,7 +12,7 @@
             <MudSpacer />
             <MudIconButton Icon="@Icons.Material.Filled.Menu" Color="Color.Inherit" Edge="Edge.Start" OnClick="@((e) => ToggleDrawer())" />
         </MudAppBar>
-        <MudDrawer @bind-Open="@_drawerOpen" Elevation="2" Anchor="Anchor.Right" Variant="@DrawerVariant.Temporary" OverlayAutoClose="true">
+        <MudDrawer @bind-Open="@DrawerOpen" Elevation="2" Anchor="Anchor.Right" Variant="@DrawerVariant.Temporary" OverlayAutoClose="true">
             <MudDrawerHeader>
                 <MudText Typo="Typo.h6">@Main_menu_string</MudText>
             </MudDrawerHeader>
