+++ src/Witness/NTS.Witness.Web/Components/Layout/NavMenu.razor
@@ -2,7 +2,7 @@
 @using Not.Blazor.Components
 @using Not.Blazor.Navigation
 
-@inherits NComponent
+@inherits NavMenuBehind
 
 <MudNavMenu>
     <AuthorizeView Roles="admin,official">
